chore: start reorganizing models/petrinets (#740)
